refactor(views): fix scroll-containment overlay positioning and preserve scrollbar heights
- Restructure Teachers, Students, Attendance, and SyncHub views to render fixed/absolute overlay modals and drawers outside the animated fade-in-up container.\n- Optimize DataTable component to keep table rows mounted during background reloads to prevent scrollbar collapse.\n- Add loading overlay on table when data is refreshing.

+++ public_shell/android/src/main/kotlin/com/nexus/school/data/StudentData.kt
@@ -12,7 +12,7 @@ import kotlinx.serialization.Serializable
 data class HonorRollItem(
     val name: String,
     val subject: String,
-    val score: Int
+    val score: Double
 )
 
 @Serializable
@@ -42,7 +42,7 @@ data class StudentScore(
     val student_id: String,
     val subject: String,
     val component_key: String, // e.g. "CA1", "CA2", "Exam"
-    val score: Int
+    val score: Double
 )
 
 @Serializable
